refactor: delete incident permission

+++ backend/src/http/controllers/delete-incident.controller.ts
@@ -1,6 +1,14 @@
 import { Authorization } from '@/auth/decorators/authorization.decorator'
+import { User } from '@/auth/decorators/user.decorator'
+import { UserPayload } from '@/auth/jwt.strategy'
 import { PrismaService } from '@/database/prisma.service'
-import { Controller, Delete, NotFoundException, Param } from '@nestjs/common'
+import {
+  Controller,
+  Delete,
+  ForbiddenException,
+  NotFoundException,
+  Param,
+} from '@nestjs/common'
 
 @Authorization('delete', 'incident')
 @Controller('/incidents/:incidentId')
@@ -8,7 +16,20 @@ export class DeleteIncidenController {
   constructor(private prisma: PrismaService) {}
 
   @Delete()
-  async execute(@Param('incidentId') incidentId: string) {
+  async execute(
+    @User() { sub: userId }: UserPayload,
+    @Param('incidentId') incidentId: string,
+  ) {
+    const user = await this.prisma.user.findUnique({
+      where: {
+        id: userId,
+      },
+    })
+
+    if (!user) {
+      throw new NotFoundException('Usuário não encontrado')
+    }
+
     const incident = await this.prisma.incident.findUnique({
       where: {
         id: incidentId,
@@ -17,6 +38,10 @@ export class DeleteIncidenController {
 
     if (!incident) {
       throw new NotFoundException('Incidente não encontrado')
+    }
+
+    if (user.role === 'CITIZEN' && incident.authorId !== userId) {
+      throw new ForbiddenException('Você não tem permissão')
     }
 
     await this.prisma.incident.delete({
